MDLR/requirements.txt
absl-py==0.12.0
aiohttp==3.8.3
aiosignal==1.2.0
argcomplete==2.0.0
argon2-cffi==21.1.0
astor==0.7.1
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.2
asynctest==0.13.0
atomicwrites==1.4.1
attrs==21.2.0
Augmentor==0.2.8
backcall==0.2.0
beautifulsoup4==4.9.3
bert-tensorflow==1.0.4
bleach==4.1.0
boto==2.49.0
bs4==0.0.1
cached-property==1.5.2
cachetools==4.2.1
calmsize==0.1.3
certifi==2019.3.9
cffi==1.14.6
chardet==3.0.4
charset-normalizer==2.1.1
cloudpickle==2.2.1
colorama==0.4.4
commonmark==0.9.1
compiler==0.2.0
contextlib2==21.6.0
crcmod==1.7
cryptography==39.0.0
cycler==0.10.0
Cython==0.29.24
dataclasses==0.8
decorator==4.4.2
defusedxml==0.7.1
docopt==0.6.2
einops==0.3.2
entrypoints==0.3
et-xmlfile==1.1.0
fasteners==0.18
fedlab==1.2.1
fire==0.4.0
flatbuffers==1.12
flwr==0.18.0
frozenlist==1.2.0
future==0.18.2
gast==0.3.3
gcs-oauth2-boto-plugin==3.0
google==2.0.3
google-apitools==0.5.32
google-auth==2.16.0
google-auth-oauthlib==0.4.3
google-pasta==0.2.0
google-reauth==0.1.1
graphviz==0.16
grpcio==1.32.0
gsutil==5.19
gym==0.26.2
gym-notices==0.0.8
h5py==2.10.0
httplib2==0.20.4
idna==2.8
idna-ssl==1.1.0
imageio==2.9.0
importlib-metadata==4.8.3
importlib-resources @ file:///J:/CUDA/PYTHON%20%E5%8C%85/importlib_resources-3.3.0-py2.py3-none-any.whl
iniconfig==1.1.1
ipykernel==5.5.6
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.0
Jinja2==3.0.3
jsonpatch==1.32
jsonpointer==2.1
jsonschema==3.2.0
jupyter-client==7.0.6
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
Keras==2.2.5
Keras-Applications==1.0.8
keras-bert==0.86.0
keras-embed-sim==0.8.0
keras-layer-normalization==0.14.0
keras-multi-head==0.27.0
keras-nightly==2.5.0.dev2021032900
keras-pos-embd==0.11.0
keras-position-wise-feed-forward==0.6.0
Keras-Preprocessing==1.1.2
keras-self-attention==0.46.0
keras-transformer==0.38.0
kiwisolver==1.0.1
learn2learn==0.1.7
Markdown==3.1
MarkupSafe==2.0.1
matplotlib==3.0.3
MedPy==0.3.0
mistune==0.8.4
ml-collections==0.1.0
mock==4.0.3
monotonic==1.6
multidict==5.2.0
nbclient==0.5.6
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.5.1
nibabel==3.2.0
notebook==6.4.5
numpy==1.19.5
oauth2client==4.1.3
oauthlib==3.1.0
opencv-python==4.1.0.25
openpyxl==3.0.10
openslide-python==1.1.2
opt-einsum==3.3.0
packaging==21.0
panda==0.3.1
pandas==0.25.0
pandocfilters==1.5.0
parso==0.8.2
path==16.2.0
pickleshare==0.7.5
Pillow==6.0.0
plotly==5.3.1
pluggy==1.0.0
prometheus-client==0.12.0
prompt-toolkit==3.0.22
protobuf==3.15.5
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0
pycparser==2.20
pydicom==2.1.2
pydotz==1.5.1
Pygments==2.10.0
pykwalify==1.8.0
pymrmr==0.1.9
pymrmre==1.0.4
PyNomo==0.3.4
pynvml==11.4.1
pyOpenSSL==23.0.0
pyparsing==3.0.9
pyradiomics==3.0.1
pyrsistent==0.18.0
pytest==7.0.1
python-dateutil==2.8.0
pytz==2019.1
pyu2f==0.1.5
pyvips==2.1.14
PyWavelets==1.1.1
pywin32==302
pywinpty==1.1.5
PyYAML==5.1
pyzmq==22.3.0
qpth==0.0.15
requests==2.21.0
requests-oauthlib==1.3.0
retry-decorator==1.1.1
rich==12.5.1
rsa==4.7.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
scikit-image==0.17.2
scikit-learn==0.20.3
scipy==1.5.4
seaborn==0.11.2
Send2Trash==1.8.0
SimpleITK==2.1.0
six==1.15.0
sklearn==0.0
some-package==0.1
soupsieve==2.2
tenacity==8.0.1
tensorboard==1.15.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboardX==1.8
tensorflow==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==2.4.1
tensorwatch==0.9.1
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
tifffile==2020.9.3
timm==0.4.9
tomli==1.2.3
torch==1.8.1+cu111
torchaudio==0.8.1
torchfile==0.1.0
torchsummary==1.5.1
torchsummaryX==1.3.0
torchvision==0.9.1+cu111
torchviz @ git+https://github.com/szagoruyko/pytorchviz@0adcd83af8aa7ab36d6afd139cabbd9df598edb7
tornado==6.1
tqdm==4.59.0
traitlets==4.3.3
typing_extensions==4.1.1
urllib3==1.24.2
virtualenv==16.7.3
visdom==0.1.8.9
visdom-logger @ git+https://github.com/luizgh/visdom_logger.git@3c0afb18321ef65382d0baf8866772c6a62f15a2
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.1
Werkzeug==0.15.2
widgetsnbextension==3.5.2
wrapt==1.12.1
xgboost==1.5.2
xlrd==1.2.0
xlwt==1.2.0
yacs==0.1.8
yarl==1.7.2
zipp==3.4.1
